Place PNLF at top-level (thanks ExtremeXT)
Detecting the existing iGPU device can be tough due to differing naming schemes, addressing, etc.  This commit alleviates a lot of the issues surrounding that as we don't need to place the PNLF device directly under the iGPU device.  Existing PNLF devices/methods/etc are still checked for and renamed to XNLF if detected to avoid potential conflicts.