Update .bat starters to work around some weirdness
Checking out the multiple-tables branch in Windows resulted in a number of PATH and ComSpec var related functions failing which manifested as python not being found, powershell not being recognized as an internal/external command, etc.  Much of this seems to be related to some scoping issues related to the line feed char that is used to replace mulitple paths erroneously set in the ComSpec var.