FakeEC Laptop updates
When working with a laptop's PNP0C09 (EC) device, qualify it (even if it's named "EC"), and ensure that if it has an _STA method we attempt to verify the return value.  If it contains multiple "Return (" entries, or just one and it's not "Return (0x0F)" - or it is an integer that does not return 0x0F, generate an _STA to XSTA rename (but keep it disabled by default), and use CondRefOf () checks to verify whether or not we can replace the method at runtime.